【Refactor】CP & SP Model Side Code Refactoring
