+++ vllm_ascend/utils.py
@@ -525,6 +525,12 @@ def register_ascend_customop():
     from vllm_ascend.ops.linear import AscendDeepseekV2MLP
     CustomOp.register_oot(_decorated_op_cls=AscendDeepseekV2MLP, name="DeepseekV2MLP")
 
+    from vllm_ascend.ops.layers.qwen3_decoder_layer import AscendQwen3DecoderLayer
+    CustomOp.register_oot(_decorated_op_cls=AscendQwen3DecoderLayer, name="Qwen3DecoderLayer")
+
+    from vllm_ascend.ops.layers.qwen3_moe_decoder_layer import AscendQwen3MoeDecoderLayer
+    CustomOp.register_oot(_decorated_op_cls=AscendQwen3MoeDecoderLayer, name="Qwen3MoeDecoderLayer")
+
     # NOTE: Keep this at last to ensure all custom actions are registered
     _ASCEND_CUSTOMOP_IS_REIGISTERED = True
